viz/oscilleditor/scripts/geometry_clip_service/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
torch>=2.0.0
transformers>=4.38.0
numpy>=1.24.0
pillow>=10.0.0
python-multipart>=0.0.9


